requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-storage
python-docx
